FE: remove init function before bootstrap and fetch config in HTTP
This also remove AppModule is not NgModule error
